requirements.txt
tensorflow==2.18.0
scikit-learn==1.4.2
streamlit==1.47.0
pandas==2.3.1
nltk==3.8.1
modal
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
python-dotenv
